refactor(rag): extract BitmapUtils + fix bitmap leak in batch enrichment
- New BitmapUtils.decodeSampledBitmap in core/data/util — single source
  of truth for the two-pass sampled decode used by both ViewModels
- SettingsViewModel: replace local decodeSampledBitmap with BitmapUtils;
  recycle bitmap in finally after captionEnrichmentProvider.describe()
- ClothingFormViewModel: replace local decodeSampledBitmap with BitmapUtils
  (recycle already in place from prior fix)
